Refactor: 트렌드 페이지 컴포넌트 분리
